+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.14.0-beta.10
+* fix import-s3-stream utility
+* provide default flow control values
+
 # 0.14.0-beta.9
 * introduce import stream utility
 * improve error reporting